requirements.txt
streamlit
pypdf2 
langchain
langchain_community
python-dotenv 
faiss-cpu 
openai 
huggingface_hub
